move url creation function to util/urlHelpers.js
