+++ webui/src/components/BannerHeader.vue
@@ -6,11 +6,7 @@
     <div class="font-bold mr-8">This is an early release</div>
     <div class="align-items-center hidden lg:flex">
       <span class="line-height-3">Please report any bugs, enhancements or features in the </span>
-      <a
-        v-ripple
-        href="https://github.com/mandiant/capa/issues"
-        class="flex align-items-center ml-2 mr-8 text-white"
-      >
+      <a v-ripple href="https://github.com/mandiant/capa/issues" class="flex align-items-center ml-2 mr-8 text-white">
         <span class="no-underline font-bold">Github issues</span>
         <i class="pi pi-github ml-2"></i>
       </a>
