+++ webui/src/components/NavBar.vue
@@ -7,6 +7,7 @@ const items = ref([
   {
     label: 'Import Analysis',
     icon: 'pi pi-file-import',
+    // TODO(s-ff): This is not the conventinal way of navigating to a new page.
     command: () => window.location.replace(window.location.origin + '/capa/') // reload the page
   }
 ])
